Move over SSH certificate management
This includes a helper `Operation#with_ssh_cmd` method that be called by
clients to generate SSH a SSH connection (i.e. a certificate) and
generate a suitable array of arguments to call SSH with.

This was originally going to be in the CLI, but it makes sense to
include it here so that other consumers can use these helpers. Namely,
Primetime will be able to take advantage of this to access deployment
logs via Dynamic Uno.

See https://github.com/aptible/api.aptible.com/pull/448 for additional
context.