Removes unnecessary dependencies
This change simply removes dependencies specified in the Gemfile,
rather than the gemspec (which seem to have been intended for
development purpose) and allows more versions of bundler to work.

The most important part is the removal of the dependency on
activesupport version 4. The libraries that this one depends on
allow both 4 and 5 to work, and this library does not appear
to actually require activesupport at all.